enh(Code): Simplify code achitecture
