+++ lib/data/repositories/cloud_benefit_imp.dart
@@ -3,9 +3,12 @@ import 'package:injectable/injectable.dart';
 import 'package:trakli/core/error/failures/failures.dart';
 import 'package:trakli/core/error/repository_error_handler.dart';
 import 'package:trakli/data/datasources/benefits/cloud_benefit_remote_data_source.dart';
-import 'package:trakli/data/mappers/cloud_benefit_mapper.dart';
 import 'package:trakli/domain/entities/cloud_benefit_entity.dart';
-import 'package:trakli/domain/repositories/cloud_benefit_repository.dart';
+
+abstract class CloudBenefitRepository {
+  Future<Either<Failure, CloudBenefitEntity>> fetchCloudBenefits();
+}
+
 
 @Singleton(as: CloudBenefitRepository)
 class CloudBenefitRepositoryImpl implements CloudBenefitRepository {
@@ -19,7 +22,7 @@ class CloudBenefitRepositoryImpl implements CloudBenefitRepository {
   Future<Either<Failure, CloudBenefitEntity>> fetchCloudBenefits() async {
     return RepositoryErrorHandler.handleApiCall(() async {
       final response = await _remoteDataSource.fetchCloudBenefits();
-      return CloudBenefitMapper.toDomain(response);
+      return response.toEntity();
     });
   }
 }
