+++ lib/data/repositories/subscription_repository_imp.dart
@@ -3,9 +3,15 @@ import 'package:injectable/injectable.dart';
 import 'package:trakli/core/error/failures/failures.dart';
 import 'package:trakli/core/error/repository_error_handler.dart';
 import 'package:trakli/data/datasources/subscription/subscription_remote_data_source.dart';
-import 'package:trakli/data/mappers/subscription_mapper.dart';
 import 'package:trakli/domain/entities/subscription_entity.dart';
-import 'package:trakli/domain/repositories/subscription_repository.dart';
+
+abstract class SubscriptionRepository {
+  Future<Either<Failure, SubscriptionEntity>> fetchSubscriptionPlans({
+    required String region,
+  });
+}
+
+
 
 @Singleton(as: SubscriptionRepository)
 class SubscriptionRepositoryImpl implements SubscriptionRepository {
@@ -23,7 +29,7 @@ class SubscriptionRepositoryImpl implements SubscriptionRepository {
       final response = await _remoteDataSource.fetchSubscriptionPlans(
         region: region,
       );
-      return SubscriptionMapper.toDomain(response);
+      return response.toEntity();
     });
   }
 }
