--- lib/domain/repositories/subscription_repository.dart
@@ -1,9 +1,0 @@
-import 'package:fpdart/fpdart.dart';
-import 'package:trakli/core/error/failures/failures.dart';
-import 'package:trakli/domain/entities/subscription_entity.dart';
-
-abstract class SubscriptionRepository {
-  Future<Either<Failure, SubscriptionEntity>> fetchSubscriptionPlans({
-    required String region,
-  });
-}
