+++ lib/domain/usecases/cloud_benefits/fetch_benefits_usecase.dart
@@ -2,9 +2,9 @@ import 'package:fpdart/fpdart.dart';
 import 'package:injectable/injectable.dart';
 import 'package:trakli/core/error/failures/failures.dart';
 import 'package:trakli/core/usecases/usecase.dart';
+import 'package:trakli/data/repositories/cloud_benefit_imp.dart';
 import 'package:trakli/domain/entities/cloud_benefit_entity.dart'
     show CloudBenefitEntity;
-import 'package:trakli/domain/repositories/cloud_benefit_repository.dart';
 
 @injectable
 class FetchBenefits implements UseCase<CloudBenefitEntity, NoParams> {
