+++ lib/domain/usecases/subscription/fetch_subscription_usecase.dart
@@ -2,8 +2,8 @@ import 'package:fpdart/fpdart.dart';
 import 'package:injectable/injectable.dart';
 import 'package:trakli/core/error/failures/failures.dart';
 import 'package:trakli/core/usecases/usecase.dart';
+import 'package:trakli/data/repositories/subscription_repository_imp.dart';
 import 'package:trakli/domain/entities/subscription_entity.dart';
-import 'package:trakli/domain/repositories/subscription_repository.dart';
 
 @injectable
 class FetchSubscriptionPlans implements UseCase<SubscriptionEntity, FetchSubscriptionUseCaseParams> {
